requirements.txt
certifi==2019.3.9
chardet==3.0.4
decorator==4.4.0
defusedxml==0.5.0
Django==2.1.7
django-geoposition==0.3.0
django-paypal==0.5.0
future==0.17.1
gmane==0.1.dev25
hello-pylab-example==0.0.1
html5lib==1.0.1
idna==2.8
lxml==4.3.2
networkx==2.2
numpy==1.16.2
oauthlib==3.0.1
parsedatetime==2.4
Pillow==5.4.1
PyJWT==1.7.1
PyPDF2==1.26.0
PySocks==1.6.8
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.2.0
scipy==1.2.1
selenium==3.141.0
six==1.12.0
social==0.1.dev3
social-auth-app-django==3.1.0
social-auth-core==3.1.0
splinter==0.10.0
stripe==2.23.0
twilio==6.25.2
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
